+++ ignatov.maxim/T3/functions.h
@@ -1,6 +1,9 @@
 #ifndef FUNCTIONS_H
 #define FUNCTIONS_H
+#include <iostream>
+#include <vector>
+
 std::vector<Point> normalize(const Polygon& poly);
-int rmecho(std::vector<Polygon>& polygons, const Polygon& target);
-int same(const std::vector<Polygon>& polygons, const Polygon& target);
+size_t rmecho(std::vector<Polygon>& polygons, const Polygon& target);
+size_t same(const std::vector<Polygon>& polygons, const Polygon& target);
 #endif /* FUNCTIONS_H */
